+++ src/js/app/reducers.js
@@ -12,7 +12,7 @@ import { medicamentoReducer } from '../medicamento/medicamentoReducer';
 import { pacienteReducer } from '../paciente/pacienteReducer';
 import { practicaReducer } from '../practica/practicaReducer';
 import { presentacionReducer } from '../presentaciones/presentacionReducer';
-import { estudiosSinPresentarReducer } from '../nueva-presentacion/estudiosSinPresentarReducer';
+import { estudiosSinPresentarReducer } from '../presentaciones/nueva-presentacion/estudiosSinPresentarReducer';
 import { loginReducer } from '../login/loginReducer';
 import { alertReducer } from '../utilities/components/alert/alertReducer';
 import { comprobantesReducer } from '../comprobantes/comprobantesReducer';
@@ -33,7 +33,7 @@ import { medicamentosEpic } from '../medicamento/medicamentoEpic';
 import { pacienteEpic } from '../paciente/pacienteEpic';
 import { practicaEpic } from '../practica/practicaEpic';
 import { presentacionEpic } from '../presentaciones/presentacionEpic';
-import { estudiosSinPresentarEpic } from '../nueva-presentacion/estudiosSinPresentarEpic';
+import { estudiosSinPresentarEpic } from '../presentaciones/nueva-presentacion/estudiosSinPresentarEpic';
 import { loginEpic } from '../login/loginEpic';
 import { comprobantesEpic } from '../comprobantes/comprobantesEpic';
 import { movimientosCajaEpic } from '../caja/cajaEpic';
