+++ src/js/presentaciones/nueva-presentacion/api.js
@@ -1,4 +1,4 @@
-import { get } from '../utilities/rest';
+import { get } from '../../utilities/rest';
 
 export function getEstudiosSinPresentarObraSocial(idObraSocial) {
     const url = `/api/obra_social/${idObraSocial}/estudios_sin_presentar`;
