+++ src/js/presentaciones/nueva-presentacion/components/NuevaPresentacionPage.js
@@ -1,12 +1,12 @@
 import React from 'react';
-import SearchPresentacionObraSocial from './SearchPresentacionesObraSocial';
+import SearchPresentacionesObraSocial from './SearchPresentacionesObraSocial';
 import TabNavigator from './TabNavigator';
 
 function NuevaPresentacionPage() {
     return (
         <div>
             <h1>Crear Nueva Presentacion</h1>
-            <SearchPresentacionObraSocial />
+            <SearchPresentacionesObraSocial />
             <TabNavigator />
         </div>
     );
